Backend/requirements.txt
Flask
langchain-google-genai
pillow
pymongo
langchain_community
fastembed
python-dotenv
